Loan cycle implement

See merge request sunilbabu/sb-solution!458

+++ kb-api/src/main/java/com/sb/solutions/api/proposal/entity/Proposal.java
@@ -30,8 +30,8 @@ public class Proposal extends BaseEntity<Long> {
     private BigDecimal proposedLimit;
     private Double tenureDurationInMonths;
     private String checkedData;
-    private Double existingLimit;
-    private Double outStandingLimit;
+    private BigDecimal existingLimit;
+    private BigDecimal outStandingLimit;
     private Double collateralRequirement;
     private String limitExpiryMethod;
     private Double duration;
